유저별 월별 데이터 조회 기능 구현

+++ app/src/main/java/com/example/cheesepay/model/TaskDTO.kt
@@ -7,10 +7,10 @@ import kotlinx.parcelize.Parcelize
 data class TaskDTO(
     var date : String = "",     //  날짜
     var name : String = "",     // 이름
-    var hourPay : Int = 0,      // 시급
-    var workHour : Int = 0,     // 근무시간
-    var extraPay : Int? = null,     // 추가금액
+    var hourPay : Long = 0,      // 시급
+    var workHour : Long = 0,     // 근무시간
+    var extraPay : Long? = null,     // 추가금액
     var extraDescription : String? = null,      // 추가금액 지급 사유
 //    var withHoldingTax : Int,     // 원천징수금액은 계산해서 구하자
-    var totalPay : Int = 0
+    var totalPay : Long = 0
 ) : Parcelable
